♻️  refactor: restructure form by separating container and component
- Converted index.jsx into form container acting as entry point
- Moved completed form UI into separate component for clarity
- Migrated payment method fetch logic from HomePage to form container
