+++ src/pages/HomePage.jsx
@@ -3,20 +3,15 @@ import { useOutletContext } from 'react-router-dom';
 import deepEqual from '../shared/utils/deepEqual';
 import getTimestamp from '../shared/utils/getTimestamp';
 import { v4 as uuidv4 } from 'uuid';
-import CATEGORY_TYPES from '../shared/constants/categoryOptions';
-import useFetchPayments from '../shared/hooks/useFetchPayments';
 import initialFormState from '../features/form/reducers/initialFormState';
 import formReducer from '../features/form/reducers/formReducer';
 import recordSchema from '../features/form/utils/recordSchema';
 import Form from '../features/form';
 import Record from '../features/record';
 
-const TEST_USER_ID = 1;
-
 function HomePage() {
   const [formData, dispatch] = useReducer(formReducer, initialFormState);
   const [originalFormData, setOriginalFormData] = useState(null);
-  const { payments, loading, error } = useFetchPayments(TEST_USER_ID);
   const { records, dispatch: recordDataDispatch } = useOutletContext();
 
   const validationResult = recordSchema.safeParse(formData);
@@ -69,17 +64,12 @@ function HomePage() {
     dispatch({ type: 'INIT_EDIT', payload: record });
   };
 
-  if (loading) return <div>결제수단을 불러오는 중입니다...</div>;
-  if (error) return <div>에러: {error}</div>;
-
   return (
     <>
       <Form
         formData={formData}
         onChange={handleChange}
         onSubmit={handleSubmit}
-        paymentOptions={payments}
-        categoryOptions={CATEGORY_TYPES[formData.type]}
         isFormValid={isSubmitEnabled}
       />
       <Record recordData={records} onSelect={handleEdit} />
